requirements.txt
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2
tabula-py==2.9.0
PyPDF2==3.0.1
pdfminer.six==20221105
Flask==3.1.3
ipykernel
ipython
keyring==25.7.0
numpy==1.26.4
scipy==1.13.1
plotly==6.3.0
matplotlib==3.10.8
requests==2.32.5
pyyaml==6.0.3
openai==1.58.1
torch>=2.0.0
tqdm==4.67.1
us>=3.2.0
kaleido>=1.0.0
beautifulsoup4>=4.12.0
